requirements.txt
openai>=1.3.7
python-dotenv
streamlit